some :lipstick: and refactorings
- move native menubar control into native title (cc @sbatten)
- introduce codeeditor.ts for reusable editor things in workbench
- inline cyclic check into core (cc @alexdima)
