+++ src/vs/code/electron-sandbox/workbench/workbench.js
@@ -50,8 +50,8 @@
 
 	/**
 	 * @returns {{
-	 *   load: (modules: string[], resultCallback: (result, configuration: object) => any, options: object) => unknown,
-	 *   globals: () => typeof import('../../../base/parts/sandbox/electron-sandbox/globals'),
+	 *   load: (modules: string[], resultCallback: (result, configuration: import('../../../platform/windows/common/windows').INativeWindowConfiguration) => any, options: object) => unknown,
+	 *   globals: () => typeof import('../../../base/parts/sandbox/electron-sandbox/globals')
 	 * }}
 	 */
 	function bootstrapWindowLib() {
