+++ src/vs/workbench/browser/parts/panel/panelActions.ts
@@ -158,7 +158,6 @@ export class PlaceHolderToggleCompositePinnedAction extends ToggleCompositePinne
 	}
 }
 
-
 export class SwitchPanelViewAction extends Action {
 
 	constructor(
