+++ src/vs/workbench/browser/parts/sidebar/media/sidebarpart.css
@@ -3,11 +3,6 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-/* Removed to allow progress bar positioning to escape */
-/* .monaco-workbench .sidebar > .content {
-	overflow: hidden;
-} */
-
 .monaco-workbench.nosidebar > .part.sidebar {
 	display: none !important;
 	visibility: hidden !important;
