+++ src/vs/workbench/browser/parts/titlebar/titlebarPart.ts
@@ -86,7 +86,7 @@ export class TitlebarPart extends Part implements ITitleService {
 		@IEditorService private readonly editorService: IEditorService,
 		@IWorkbenchEnvironmentService protected readonly environmentService: IWorkbenchEnvironmentService,
 		@IWorkspaceContextService private readonly contextService: IWorkspaceContextService,
-		@IInstantiationService private readonly instantiationService: IInstantiationService,
+		@IInstantiationService protected readonly instantiationService: IInstantiationService,
 		@IThemeService themeService: IThemeService,
 		@ILabelService private readonly labelService: ILabelService,
 		@IStorageService storageService: IStorageService,
