+++ src/vs/workbench/common/editor/textDiffEditorModel.ts
@@ -14,14 +14,14 @@ import { DiffEditorModel } from 'vs/workbench/common/editor/diffEditorModel';
  */
 export class TextDiffEditorModel extends DiffEditorModel {
 
-	protected readonly _originalModel: BaseTextEditorModel | null;
-	get originalModel(): BaseTextEditorModel | null { return this._originalModel; }
+	protected readonly _originalModel: BaseTextEditorModel | undefined;
+	get originalModel(): BaseTextEditorModel | undefined { return this._originalModel; }
 
-	protected readonly _modifiedModel: BaseTextEditorModel | null;
-	get modifiedModel(): BaseTextEditorModel | null { return this._modifiedModel; }
+	protected readonly _modifiedModel: BaseTextEditorModel | undefined;
+	get modifiedModel(): BaseTextEditorModel | undefined { return this._modifiedModel; }
 
-	private _textDiffEditorModel: IDiffEditorModel | null = null;
-	get textDiffEditorModel(): IDiffEditorModel | null { return this._textDiffEditorModel; }
+	private _textDiffEditorModel: IDiffEditorModel | undefined = undefined;
+	get textDiffEditorModel(): IDiffEditorModel | undefined { return this._textDiffEditorModel; }
 
 	constructor(originalModel: BaseTextEditorModel, modifiedModel: BaseTextEditorModel) {
 		super(originalModel, modifiedModel);
@@ -73,7 +73,7 @@ export class TextDiffEditorModel extends DiffEditorModel {
 		// inside. We never created the two models (original and modified) so we can not dispose
 		// them without sideeffects. Rather rely on the models getting disposed when their related
 		// inputs get disposed from the diffEditorInput.
-		this._textDiffEditorModel = null;
+		this._textDiffEditorModel = undefined;
 
 		super.dispose();
 	}
