+++ src/vs/workbench/contrib/debug/browser/callStackView.ts
@@ -33,7 +33,6 @@ import { ActionBar } from 'vs/base/browser/ui/actionbar/actionbar';
 import { isSessionAttach } from 'vs/workbench/contrib/debug/common/debugUtils';
 import { STOP_ID, STOP_LABEL, DISCONNECT_ID, DISCONNECT_LABEL, RESTART_SESSION_ID, RESTART_LABEL, STEP_OVER_ID, STEP_OVER_LABEL, STEP_INTO_LABEL, STEP_INTO_ID, STEP_OUT_LABEL, STEP_OUT_ID, PAUSE_ID, PAUSE_LABEL, CONTINUE_ID, CONTINUE_LABEL } from 'vs/workbench/contrib/debug/browser/debugCommands';
 import { ICommandService } from 'vs/platform/commands/common/commands';
-import { CollapseAction } from 'vs/workbench/browser/viewlet';
 import { IViewDescriptorService } from 'vs/workbench/common/views';
 import { textLinkForeground } from 'vs/platform/theme/common/colorRegistry';
 import { IThemeService, ThemeIcon } from 'vs/platform/theme/common/themeService';
@@ -207,7 +206,7 @@ export class CallStackView extends ViewPane {
 
 	getActions(): IAction[] {
 		if (this.stateMessage.hidden) {
-			return [new CollapseAction(() => this.tree, true, 'explorer-action ' + ThemeIcon.asClassName(icons.debugCollapseAll))];
+			return [new Action('debug.callStack.collapseAll', nls.localize('collapse', "Collapse All"), `explorer-action ${ThemeIcon.asClassName(icons.debugCollapseAll)}`, true, async () => this.tree.collapseAll())];
 		}
 
 		return [];
