+++ src/vs/workbench/electron-sandbox/desktop.contribution.ts
@@ -22,9 +22,6 @@ import { INativeHostService } from 'vs/platform/native/electron-sandbox/native';
 import { IJSONContributionRegistry, Extensions as JSONExtensions } from 'vs/platform/jsonschemas/common/jsonContributionRegistry';
 import product from 'vs/platform/product/common/product';
 import { IJSONSchema } from 'vs/base/common/jsonSchema';
-import { Extensions as WorkbenchExtensions, IWorkbenchContributionsRegistry } from 'vs/workbench/common/contributions';
-import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle';
-import { LoaderCyclicChecker } from 'vs/workbench/electron-sandbox/loaderCyclicChecker';
 
 // Actions
 (function registerActions(): void {
@@ -100,9 +97,6 @@ import { LoaderCyclicChecker } from 'vs/workbench/electron-sandbox/loaderCyclicC
 			primary: KeyMod.CtrlCmd | KeyMod.Shift | KeyCode.KEY_I,
 			mac: { primary: KeyMod.CtrlCmd | KeyMod.Alt | KeyCode.KEY_I }
 		});
-
-		Registry.as<IWorkbenchContributionsRegistry>(WorkbenchExtensions.Workbench).registerWorkbenchContribution(LoaderCyclicChecker, LifecyclePhase.Ready);
-
 	})();
 
 	// Actions: Runtime Arguments
