+++ src/vs/workbench/electron-sandbox/parts/dialogs/dialogHandler.ts
@@ -206,4 +206,3 @@ export class NativeDialogHandler implements IDialogHandler {
 		}
 	}
 }
-
